Verify password, adjustements for LDAP
In case LDAP is active, dont show the password change form, and print
the warning message explaining that passwords can't be modified.

When LDAP login is active, or in a genereal case, the user cannot change
its own password, this page should not be reached in the first place,
because the options to reset user password should not be available.