Include realname field in verify.php
Show realname field in account verification, and allow the user to
modify their real name in the same page.

This is useful for self registration, where real name is created blank.
So even if real name is not required, there is more chances that the
user fills in their name on first activation.