Add email notifications for users assigned to categories
Users who are assigned to categories can now receive email
notifications. This is configurable via E-mail Notifications page,
on a per-project basis.

A common use case:
- A user was assigned to some category;
- When new issue is submitted to this category and "Assigned To" field
  is empty, this issue is assigned to Category Owner (who receives a
  notification);
- If we then re-assign issue to someone else, the Category Owner will no
  longer receive notifications.

This commit makes it possible for category owners receiving
notifications in this case as well.

Fixes #20876

Signed-off-by: Damien Regad <dregad@mantisbt.org>

Original commit amended to default the new flag to ON as per [1].
Reworded and formatted commit message.

[1] https://github.com/mantisbt/mantisbt/pull/762#issuecomment-217280123