Implement database api functions for sql IN clauses
Implement db_sql_in() function in database_api to ease the creation
of long list of values, to be used as part of IN clauses.
This function makes use of DBTempData class, so if the list is bigger
than a defined size, it loads the data into a temporary table and
returns the equivalent select query.
This conversion transparent to the caller, and can solve some problems
with query length limits.