Don't cache users in project_get_all_user_rows()
Most of the callers of project_get_all_user_rows use the return array
directly, and do not need to fetch the user rows. There's no need to
always cache the calculated users. Especially when the user list is very
large, eg: print users option list with public access.  By not caching
them, memory and time is saved (see issue #17577)

Of all the callers of this function, those that actually use the
cached data are:
 - email_notify_new_account(): Adds an explicit call to cache
 - email_collect_recipients(): Was already calling to cache users

Signed-off-by: Damien Regad <dregad@mantisbt.org>