Hide DB2 for new installs
The MantisBT team have no access to a DB2 environment and no technical
knowledge of this database. Consequently, we are not able to support it
moving forward.

This commit removes 'db2' from the database types selection list in the
installer, reflecting the fact that this DB type is deprecated.

This commit should actually have gone in 1.3.0, but I forgot to merge
the branch back then...

Fixes #17336