Do not display the access level in bugnotes when it is 0 (ANYBODY)
Prior to this, the access level display in bugnotes was (@0@) if the
user who added the note did not have access to the project. This
situation can happen when an issue is moved to a private project, or
when a user's rights to a private project are revoked.
This is similar to the fix for #11923, but affects email notification.

Fixes #20897