Do not add a space before version suffix
It makes more sense to let the user choose whether they need a space
before the suffix by adding it themselves, rather than hardcoding it
in core which prevents them from removing it.

This reverts commit 065521c1f7832d12fd0b4ca47e7587caa214444b and applies
the same change in manage_overview_page.php. The admin index page did
not need fixing as it already concatenated the suffix without the space.

Fixes #20909