Travis: only install publican for Docbook builds
We do not need Publican for regular MantisBT builds.

This commit improves the overall build performance by specifying the
addon apt package install in the build matrix for the docbook specific
build only.

The average gain is approx 30-45 seconds per build, for a total improvement of
over 5 minutes for all builds:
- Before: 23'25" - https://travis-ci.org/dregad/mantisbt/builds/130372109
- After:  17'55" - https://travis-ci.org/dregad/mantisbt/builds/130374454

Fixes #20910