Enhance access level display
Prior to this, the access level display in bugnotes was (@0@) if the
user who added the note did not have access to the project. This
situation can happen when an issue is moved to a private project, or
when a user's rights to a private project are revoked.
This is similar to the fix for #11923, but
- affects also email notification
- displays "no access" for existing users instead of displaying nothing

Fixes #20897