Let Timeline handle non-existing bugs
If an history entry refers to a bug that does not exist in the database,
history_get_event_from_row() throws application error 1100.

Even though it is not a normal situation to find orphan records in the
history table, the overhead of verifying a bug's existence at the
beginning of the loop is negligible, so it doesn't hurt to add the extra
bug_exists() check.

Fixes #20727