Enhance logging to show debug email when used
Prior to this, when using an invalid address, user would not see a clear
indication in the log file that the debug email was being used,
resulting in confusion caused by the error message generated by
email_send().

This commit adds a LOG_EMAIL_VERBOSE entry clearly showing when the
debug email address is used.

Fixes #20679