Use specific regex for parsing mentions
It turns out that the current $g_user_login_valid_regex pattern
cannot be used as it allows spaces. Furthermore, special handling is
required to process e-mail address-like strings.

For this reason, a custom regex is built for the purpose of mentions
parsing, supporting only a subset of the allowed usernames.

Fixes #21083