Improve handling of empty values in adm_config_set.php
When processing an empty value, the code now evaluates the config type
as if user had selected 'default'; this ensures the proper case of
'empty' is used depending on the config being set (i.e. 0, '').

We also avoid calling the config parser in this case, because a simple
typecast is sufficient to process user input.

Fixes #21124