Don't show version release date if not set
In roadmap, dont show release date if it's not set.
When the date is not explicitly set, its stored as 1 in database.
Logically it means "date null", but it was being shown as an actual
date: 1970-01-01

When not showing date, also becasue of option $g_show_roadmap_dates set
to OFF, do not show the calendar icon.

Fixes: #12409