Create BugData callback queue
A callback queue in BugData object.
The registered callbacks are performed in the update() function,
after the data are saved, and before the EVENT_UPDATE_BUG event is
called.
These callbacks are needed becasuse with some bug actions, some auxiliary
operations must be performed between the events EVENT_UPDATE_BUG_DATA
and EVENT_UPDATE_BUG.
For example: bug notes, must be inserted only if the action is
performed, which can be stopped if the validation fails.