Fix comparing threshold as an integer
report_bug_threshold may be an integer or an array, but a check to
see if an access level is "greater than" its value, was assuming
that this threshold was an integer value.

A new function is created to get an integer representation of a
threshold, as the minimum integer value defined in the threshold.

Fixes: #21579