Fix Workflow config to acount for complex thresholds
In workflow transitions configuration page, the "Minimum Access Level to
Change to this Status", for bug submitted status, is linked to the
report_bug_threshold option.
If the threshold is configured as an array, the simple dropdown select
cant be used to represent or modify this value, so don't allow it to be
editable.

Fixes: #21581