Use returned object from event instead of $this
Use the returned BugData object from the event call to comply with the
definition of EVENT_UPDATE_BUG_DATA as EVENT_TYPE_EXECUTE

The argumentation, as explained in the pull request:
>>>
Currently we have EVENT_UPDATE_BUG_DATA defined as "chain". This means
that the returned value by the event call is the object expected to be
used.

In the case of modifying a BugData object, this is not needed, as the
BugData object passed as parameter to the event is a reference, and
changes made through this reference are made into the actual (original)
object.

Let's see the BugData::update() function after this PR:

inside update(), we call:
event_signal( 'EVENT_UPDATE_BUG_DATA', $this, $t_old_data );
and later, we update the bug with the values from $this.

If the plugin makes changes to the supplied BugData reference, there is
no problem, they are saved in the actual object.
But, being a chain event, the returned reference may be altered, to
another object not being the original one. This doesnt make much sense,
but is possible though. For example:

function plugin_bug_data_hook( $event, $updated_bug, $old_bug ){
$another_data = new BugData();
...
return $another_data;
}

back to the BugData::update(), we cannot update "self" with another
reference, from inside the object.
<<<

The described exceptions are corner cases that probably should not be
supported, but as current event functionality allows it, the changes in
this commit are made to account for it.