Revert multiple tokens for attachement delete links
Revert the logic introduced by fix for Issue #20142, in which one single
form security token is used for all action links.
The side effect is that only one of those actions can be performed from
the specific originating page.
This is especially a drawback when the actions are links that can be open
in new browsers tabs.

Reverting to separate tokens allows to perform all the actions without
having to reload the originating page.

Note, the original performance issue was solved at the session layer
with 8092c3d.

Fixes: #21610