Fix managing "hide status" for advanced filters
"Hide status" field is used only in simple filters, when the filter is
switched to advanced, this field must be ignored.
When the filter is of type advanced, that field must be initialized to
"none".

Fix how the "hide status" is initialized when requesting a default
filter, by adding the parameter view_type to filter_get_default_array().
Now "hidden status" is initialized correctly according to the requested
view state.