Fix retrieving the updated version of a sotred filter
The filter active for current user is stored as an unnamed filter. But
when it was loaded from a stored "named" filter, we keep track of the
source_query_id, to be able to show in filter box, the name of that
stored filter that is in use.
But, now that stored filters can be edited, this user cached version of
filter may differ if the original named filter is modified.

Now, when loading the current active filter, if there is a
source_query_id referencig a filter id, load those filter values
instead.