Change view issue menu to allow plugin links
Changed the style of the upper menu in bug view page, to allow the
correct formatting of links provided by plugins with EVENT_MENU_ISSUE.

This event may return a prepared html link, which was being shown as a
bracketed plain link (old style)
The new style introduced by new ui, used link buttons, but the function
provided to print abutton from the event return (print_small_button),
must provide a separated urn and label, which is not possible with an
already prepared html link.

The changed style now uses list items that formats the link html "as
is" in an uniform way with the ui.

Fixes: #21559