Fix getting values from unchecked checkboxes
A checkbox input that is not checked is not submitted with the form.
In that case, while reading gpc_get of that input, we don't know if the
field was not part of the form, or the checkbox was unmarked explicitly
by the user.
The workaround is to prepend any checkbox with a hidden field that sets
the fallback value when the main box is unchecked (and not submitted).
By having the same imput name, the value submitted when names conflict is the last
successful control in the form.