Drop duplicated lang strings
There were two lang strings in use for filter filed headers. Eg:
$s_category = 'Category';
$s_category_label = 'Category:';

As the xxx_label types are the same as the simple one, which are the
general strings in use through the application, we drop the duplicated
xxx_label ones.
The ':' character must be discarded then, as semantically, it should
not be contained in the text itself. Some filter header strings have
been modified to reflect this.

Additionally, change the texts for both filter dates, as the new
functionality need them to be more focused on the actual behaviour.