Require mbstring extension
The installer will fail with 500 errors on login page if the mbstring
extension is not present, unless it can fall back on the multi-byte
functions provided by the utf8 library which requires the utf8_encode()
function (provided by the XMLParser extension).

The Admin Guide has been updated to clearly document the fact that
mbstring is mandatory.

Based on original commit proposed by Gianluca Sforna in
https://github.com/mantisbt/mantisbt/pull/903

Fixes #21959