Fix false positive for FilterTest::testGetProjectClosedIssues
The test did not work as expected (returning false successful results),
due to incorrect initialization of 'status' and 'resolution' fields for
the created test issue, resulting in status defaulting to 'new' instead
of being set to 'closed' as expected.

Note: this will cause unit tests to fail when $g_hide_status_default is
set to hide closed issues (as it is by default), see issue #21996.

Fixes #22001