Initialize timezone early
log_event() uses date(), which triggers a PHP notice when the default
timezone has not been set in php.ini (date.timezone) or by the
application.

Since core.php initializes the timezone only after including several
APIs, depending on log and error level settings in config_inc.php, this
can prevent MantisBT from operating properly.

To prevent this problem, we now set the default timezone as early as we
can, and change it later if necessary according to the user's
preferences.

Fixes #22011