Fix case for date_submitted > last_updated
In mantisbt.org's bugtracker (and probably in other ancient trackers
running since the early days of MantisBT as well), there are issues
having a creation date newer than the last_updated date.

It is not confirmed whether this is a glitch, or a consequence of the
software's evolution. To avoid incorrect display in the Issue Trend
graph, we apply a simple fix while building the data set.

Fixes #22065