Use MantisCore plugin version for dependency check
Until now, the maximum dependency check for Mantis core version in
plugin_dependency() function relied on MANTIS_VERSION constant.

While the the MantisCorePlugin::version property has by definition the
same value as the constant, using the latter prevents simulating use of
the API with a different Mantis version, which is required for unit
tests.