Expand image attachments by default
If an attachment of type image is smaller than the maximum preview
threshold ($g_preview_attachments_inline_max_size), Mantis was only
displaying the attachment's name and a [+] button, forcing an extra
click for the user to view it.

We now display images by default, allowing the user to optionally hide
them instead by clicking the [-] button.

Fixes #21796