Show "GOOD" status for installer DB connect checks
Since removal of DB2 support (see issue #0021140), the installer does
not display "GOOD" status anymore for Admin and User DB connection
checks.

This also improves the logic for the Admin account connection check, we
now capture a "BAD" status if the Admin can connect to the DB, but does
not have access to the database.

Fixes #22400