Use CF display value when logging history event
custom_field_set_value() was calling history_log_event_direct(),
passing it the old field's display value, and the new field value in
database format.

Depending on the custom field's type (only some of them actually define
a database_to_value function), this could cause a change to be logged
in history, when the field's value was in fact not modified.

We now consistently use the "display" value for both old and new value
when logging custom fields history.

Fixes #22408