Travis: force use of PHPUnit 5.7 with PHP >= 5.6
On 2017-03-09, Travis deployed the latest version of PHPUnit (6.0.8) on
their workers. Since that day, our builds on PHP >= 5.6 are failing,
because 6.0.8 only supports PHP 7.0 and 7.1 (<= 5.5 are OK, because they
are running an older PHPUnit version).

PHP 7.x builds are failing as well, because PHPUnit 6.x introduced
changes that are breaking the MantisBT tests.

As a workaround, we download and install PHPUnit 5.7 if the Travis PHP
version is 5.6 or later.

Fixes #22503