Modify schema to get install/upgrade work with db-mssql.
Although the problem appears in step 209, the real cause is the
combination of steps 200 and 201. The Index ('user_id','name')
created in step 201 (with a nullable 'user_id' as of step 200)
forbids the alteration of 'user_id' to NOTNULL in step 209, at
least with mssql. So fix this in step 200 and set 'user_id'
NOTNULL right from start.

While this is a 'post release' change, it is justifiable as it
happens in the same install/upgrade sequence and the final
db (schema) is identical.

Fixes #22063

Signed-off-by: Damien Regad <dregad@mantisbt.org>

Backported from master 3a0706a1bc291be6ee684bd30199bd6544f0cc6b