Remove functionality: serialized filter as parameter
current_user_get_bug_filter was reading for a gpc variable "filter",
which can represents a json encoded filter array
This means that this parameter would override the usual filter workflows
if present.
AFAIK, this is not used currently in mantis, and may be an old
implementation.

Keep the existing "filter" parameter as a numerical id, whcih point to
the token_id which holds a temporary filter