Remove PASSWORD_MAX_SIZE_BCRYPT constant
Considering that the CRYPT_BLOWFISH hashing algorithm that we now use
as default to encrypt passwords truncates the the hash at 72 chars [1],
we must limit the users' input accordingly.

Allow 1024 chars for the password is not really useful (the value was
picked arbitrarily), so we now use 72 as the maximum password length
for all password hashes (PASSWORD_MAX_SIZE_BEFORE_HASH).

[1] http://php.net/function.crypt