Fix highlighting of 'Manage Configuration' menu
Issue #19992 changed the default page for 'Manage Configuration' menu.
This introduced a regression, causing the main menu item to no longer be
properly highlighted when any page other than the previous default one
(Configuration Report) was active.

We now define the default page as a constant (DEFAULT_CONFIG_PAGE),
which is used in all of the menu's sub-pages.

Fixes #22730