Exclude unlinked projects from custom fields filter
Don't get data from custom field values from issues in projects that are
not linked to the custom field.
This data can exists if the field has been unlinked after data has been
saved. When unlinking a field from a project, the stored values are not
deleted, however they must be ignored.

Fixes: #19385