Fix oracle sql syntax for table alias
Fix the oracle syntax to not remove those "AS" that are part of CAST
expressions.

Oracle doesn't suppor AS keyword as part of table aliases.
Previous code was modifying the query sql to remove all appearances of
"AS". However, "AS" can also appear as part of a CAST expression:
CAST( x AS y )

Fixes: #23268