Fix testGetIssuesForUserMonitoredNoTargetUser
This was testing a filter with "monitoring user = meta_filter_none",
which was previously unsupported. That resulted in issues monitored by
user -2, so before, and after adding a new issue, the filter returns 0
issues always.

Now that meta_filter_none is supported, meaning issues which are not
monitored by any user, after adding a new issue, the filter must return
1 more issue.