Fix handling custom date fields with value `{now}`
The issue should be created with custom field set to today's date,
but it gets created with literal string {now} that can't be displayed
later. The fix would be resolve {now} at reporting time and fix display
of existing issues by showing empty date.

This applies to issues created via web ui when the date field is not
visible on reporting page.

Fixes #23594