Remove error handler's "errcontext" parameter
The $errcontext parameter for the error handler function has been
deprecated in PHP 7.2.

We currently use this parameter to display some of the local variables
when an error is triggered and $g_show_detailed_errors = ON, but the
output is broken and difficult to read since 2.x was released.

Considering the deprecation notice, it does not make sense to fix the
layout now since the raw data will no longer be available in the next
major PHP release.

This commit changes the error handler's signature, and removes all code
dedicated to printing Context information on the detailed error page.

Fixes #23942