Send always usernames in email notifications
Don't send realnames if $g_show_realname = ON;

This is a quick workaround as a clean solution needs some redesign.

At the moment $g_show_user_realname_threshold is considered based
on the current user.
This is wrong, as the option must be considered based on the
recipient of the notification.

Fixes #24239