Use prepare_user_name() to display @user mentions
mention_format_text() was refactored and simplified as the logic to
display the user's name is now handled by prepare_user_name().

prepare_user_name() was modified to have existing but disabled users
displayed as a link with strike-through (before, the username was
already wrapped in a <del> tag, but there was no link).