Fix filter by date for the soap api
In the soap endpoint “mc_filter_search_issues”, it is possible to
filter the search result by date. At least it is defined in the wsdl.
But as described in the issue #20101, this parameter has no
effect (yet).

Therefore handle the date parameters manually in the mc_filter_api,
so that it will be appended to the filter.

After this change it is possible to filter the search result by a
date range.

Fixes #20101