Obsolete option default_manage_user_prefix
The option has been introduced in 2003 to solve performance issues on
installations with thousands of users.

The option was not documented in Admin Guide.

Paging (hardcoded 50 records) has been implemented in 2008.
After this, there is no longer the need for the option.
The performance is fine, even with more than 35k users.

Issue #24689