Proper fix for duplicate 'sidebar' div ID
The initial attempt at fixing this issue [1] introduced a regression in
the persistence of the sidebar's collapse state.

Using a different ID for the toggle button or the sidebar div caused
other issues (e.g. malfunctioning hambuger menu on narrow screens, or
failure to save collapse state).

The root cause was a bug in the javascript code saving the sidebar's
state, which was referencing the toggle button's id instead of the
sidebar div's.

We now have ID 'sidebar' for the actual sidebar, and 'sidebar-btn' for
the toggle button, and the javascript has been modified accordingly.

Fixes #24976

[1]: commit b7b914549fe8fbdf4ee4053f475031df9d6a4674