Reject files in client side Dropzone
Catch the event for file added and check the file size to be under the
maximum allowed size for uploads. If file is bigger, a warning message
is shown and the file is rejected.

Change how the max_file_size is passed to client, so that now we have
the exact value in bytes to compare, even though Dropzone will still
be configured with an aproximation in MiB.

Fixes: #25464