Fix My View Page missing boxes
When displaying boxes on 2 columns (i.e. when Timeline is not shown),
the code handling the display of boxes failed to print the first box
after starting the 2nd column.

This commit simplifies the logic, only handling the closing of first
column's div and opening of 2nd column, and always printing the box
regardless of status.

It also removes dead code: the elseif to close the div at end of 2nd
column was never called, the div is actually closed after the foreach
loop ends.

Fixes #22104