Remove hidden My View boxes before displaying them
Prior to this, the hiding and display of boxes was performed within a
single foreach loop. As a consequence, depending on system configuration
and for low-privileged users, there could be an uneven distribution of
boxes between the 2 columns when Timeline is switched off.

To fix this issue, the processing of boxes is now done in 2 steps :
  1. Filtering boxes
  2. Display

Fixes #25679