Fix mssql summary error
When computing SUM, AVG, functions, sqlserver uses the column
datatype, which is INT. This datatype can be overflowed when a big
number of issues are included, since we are adding the total number
of seconds.
In that case, use an explicit cast to BIGINT to avoid that error.
Other databases are not affected becasue they use a runtime data
type allowing bigger integers.

Fixes: #25781