Fix invalid HTML in manage_config_workflow_page
The generated markup has a form to revert the configuration changes
defined within the main form. This is invalid HTML.

To fix this without complicating the layout, we use a deported <button>
within the main form, using the HTML5 `form` attribute pointing to the
the revert form's id, which is defined outside.

Fixes #25784