Let users add monitors even if they can't see them
In view.php, the 'Users monitoring this issue' box is only displayed
if the user has $g_show_monitor_list_threshold. If the same user had
$g_monitor_add_others_bug_threshold, they would not be able to adding
new monitors.

This is an uncommon situation, as monitor_add_others_bug_threshold is
expected to be >= show_monitor_list_threshold. Still it makes sense to
fix that since the configuration allows such setup.

With this change, an 'Access Denied' message is displayed instead of the
users' list if access level is below show_monitor_list_threshold and
>= monitor_add_others_bug_threshold, and the form to add new monitors is
shown.

If user's access level is lower than both thresholds, then the behavior
is unchanged, i.e. the whole box is hidden.

Fixes #25815