Revert maintenance mode in TestConfig.php
For an unknown reason, commit 59970d8735dd06975c3aeabbd9ce8ba0d7593b16
(part of issue #20837) set MANTIS_MAINTENANCE_MODE.

As a consequence, it is not possible to use any database function as
as part of unit tests.

Since the tests added in the above-mentioned commit are also successful
without maintenance mode, it makes sense to revert this change.