PR fix 'limit_reporters' filter query
When the old option 'limit_reporters' is enabled, create the correct
project clauses.
Also, since this option was only available at ALL_PROJECTS level, proect
overrides are not expected to exist (and weren't supported to work
properly). Then, force the reading of this option with config_get from
ALL_PROJECTS to simplify the logic.