Utility function to process event results
Encapsulate a common function to process the results of an event of
type "default". Usually, extracting the data from the event responses
requires a nested loop that is not simple.
This new function returns a simple array of the items extracted from the
event result.
An optional parameter for a callback function is provided to cover
complex scenarios.

Refactor the code related to EVENT_TYPE_DEFAULT results processing.