Fix circular dependencies in Relationship Graph
In the case where we have a cycle in parent-child relationships (i.e.
A -> B -> C -> A), the Relationship Graph fails with an INTERNAL
APPLICATION ERROR - Argument 1 passed to relgraph_add_child() must be of
type array, string given in core/relationship_graph_api.php line 427.

The recursive call was passing the bug id instead of the bug list.

Fixes #11381