Install: fix default prefix change by DB type
The JavaScript that changes the default table prefix/suffix based on
database type and populates the sample table name fields was broken
because multiple DOM elements with the same id existed on the page
(2 sets of hidden <span> for the default values, and the 3 <input>).

Changing the markup for the <span> to set a class instead of an id, and
ajusting the script fixed the problem.

Fixes #26596