Specific notification for password reset by admin
The email notification sent when password is reset by an administrator
was misleading, implying that it can be ignored when in fact it should
not as the password has effectively been changed to a random one. If the
user does not reset their password, they will not be able to login.

This commit avoids confusion caused by the misleading message, by
sending a specific notification for password resets by admin.

A new $p_reset_by_admin optional parameter, defaulting to false, was
added to email_send_confirm_hash_url() function, allowing it to be used
for resets by user (lost password feature) also.

Fixes #26884