Print lang string if English localization missing
When an English language string is missing, MantisBT currently displays
an empty string; this is not very friendly to the user, who may be
missing important information (e.g. a button label).

If English string is missing, display the language string itself
(i.e. the $p_string parameter to lang_get()).

Fixes #27241