Special handling for missing lang string errors
If the error type is 300 (ERROR_LANG_STRING_NOT_FOUND), we retrieve the
details (file name and line number) of the parent error. Properly
handles the case when user's language is not English, by going through
the error stack trace until the first lang_get() call is found.

This allows printing information about the actual location of the
missing string, instead of a not-so-useful reference to lang_get().

Fixes #27241