Change icons used in Manage User Page
The $t_lock_image variable used to display the padlock icon to identify
protected accounts includes a 'Protected' pop-up, so reusing it for
locked account is not possible.

This commit changes the icon used for protected accounts to a shield
(fa-shield), and defines a new variable for protected accounts with
the padlock (fa-lock) and a specific pop-up.

Fixes #36841