UserResetPasswordCommand: add 'action' parameter
Previsouly, the action (RESET or UNLOCK) was determined based on the
user account's state.

The action is now specified when instantiating the action, and defaults
to RESET; value is stored in a class property.

The constants that were used to return the action performed by the
command, have been renamed to remove the RESULT prefix, since they are
now also used to specify the action parameter.

In manage_user_edit_page.php, use named HTML buttons instead of inputs
to specify the action to execute in manage_user_reset.php.

Issue #36842