Remove array_reverse to display bug revisions
The same change was previously applied to bugnotes and revisions (see
Issue #14273, commit ce6d92de07fc370c5e69dce4794f8a60d180e6a2), but not
to bugs.

This causes the revisions to be listed in a different order when viewing
a bugs by bug_id vs by rev_id, even though it is the same data. For
consistency, the change needs to be applied everywhere.