Hide 'View Revisions' if user has no access
Do not display the 'View Revisions' link to the user if they are not
allowed to see it, both when displaying bugnotes and in bug history.

Added $p_bug_id parameter to history_localize_item() function, it is
needed to call access_can_view_revisions().

Adapted mci_issue_get_history() and email_format_bug_message() to use
the new function signature.

Fixes #20690