Prevent setting category not belonging to project
When retrieving a category for a given project, make sure that it is
available in the project's hierarchy, taking inheritance into account.

This is a follow-up on commit b77859901050b558bfcd28050cff1599d60e45fa
which only covered bug_report.php, when in fact the same problem was
also present in bug_update.php.

Fixes #27361