Properly check access level when updating issues
Commit bf1f02f71a85bd4e05cd2a627b3c25e79f96d5a0 (issues #21393, #22093)
introduced a bug in the access level check, as it passed the threshold
as a string to access_ensure_bug_level() instead of its actual value.

This seemed to work in PHP < 8.0, because string comparisons behaved
differently [[1]], but in reality there was no actual access check, as
the strings were evaluated to 0 (ANYBODY).

Adding the missing config_get() call to fix the access level check.

Fixes #27806

[1]: https://www.php.net/releases/8.0/en.php#consistent-type-errors-for-internal-functions