Validate date custom fields default value format
Prior to saving the custom field's definition, we try to convert the
provided format to a date, and if that fails an error is thrown.

This also removes the need to enclose the format within curly brackets
(e.g. {tomorrow}); this is more user-friendly, and simplifies the code
as the given value can be passed on as-is to DateTime transformation.

Backwards-compatibility is maintained for now, but may be removed in a
future release.

Fixes #27950, #27956, #27983