CSS: fix font size for labels in table cells
HTML inputs should be associated with a label, for screen readers.

A lot of our forms are in tables, with one column containing the label
and another the corresponding input.

Adding the <label> element in this scenario causes the display to
change the font size, due to the default defined in ace.css as 14px,
but in table cells it is 13px (derived from body).

This overrides the font-size for th/td labels, so the labels visual
aspect stays the same after adding <labels> elements.

Follow-up on d27ea37de79604f9e6cbbd908383085f4a5f679a.

Fixes #27958