Custom Field names with commas (`,`) are forbidden
Until now, the custom fields management page allowed saving a Custom
Field with commas; attempting to use such a field in the Manage Columns
pages results in APPLICATION ERROR 2601 when trying to save the
configuration.

This commit prevents saving or updating a custom field if its name
contains a comma.

Fixes #26665