Version field is now required
When adding a new version from manage_proj_edit_page.php via the
'Add Version' or 'Add and Edit Version', it should not be allowed to
submit an empty field.

The HTML required attribute has been added to the 'version' input, and
the manage_proj_ver_add.php page displays an error message if no version
was created.

Fixes #27994