Fix security token field name for plugin pages
When print_form_button() was called with a plugin page (as generated by
the plugin_page() API function) as p_action_page parameter, the form
security field's name is not valid.

We now detect whether $p_action_page is a plugin page (i.e. script name
is 'plugin.php') and if so generate a valid security token name from the
'page' parameter.

Fixes #28533