Button to edit custom fields in manage project page
The use of a hyperlink to edit Custom Fields is not consistent with the
other sections on the page (subprojects, versions).

This commit removes the hyperlink on the Custom Field's name, and adds
a new 'Edit' button in the Actions column instead.

We also modify the existing print_form_button() call that removes custom
field association, to use the function's $p_args_to_post parameter
instead of building a GET URL.

Fixes #28557