Use filter key instead of URL to build permalink
Refactor permalink_page.php to accept a temporary filter key and
generate the URL from that, instead of receiving a fully-formed URL.

This prevents issues when the filter criteria contain a `&` (e.g. a
category named "a & b"), causing the value to be interpreted as 2
distinct parameters due to string_sanitize_url() decoding the `%26`
before processing the query string.

Fixes #30204