Consistently set type class for custom field columns
'cftype-textarea' class was introduced to fix wrapping for textarea
custom fields, as part of the patch for the regression on overflow-wrap
(see #30268).

Assign a 'cftype-xxx' class to Custom Field columns of any type, where
xxx is the type (label) defined in $g_custom_field_type_enum_string,
with a simple transformation to ensure it is a valid CSS identifier.

Issue #30279