Use HTMLPurifier to check valid tags in lang strings
This is a first, limited attempt at using the HTMLPurifier library,
to improve the quality and security of language strings by detecting
whether any unallowed HTML tags or properties are used.

For now this is only a manual check, limited to the test_langs.php
script which will report any exceptions, but eventually we should
generalize this to sanitize all user input.

Issue #12242