Do not empty email, realname when updating with LDAP
When using LDAP with $g_use_ldap_email = ON or $g_use_ldap_realname = ON
and updating a user's record from manage_user_edit_page.php, the email
or realname fields were set to blank if the LDAP connection failed or
the user was not found in the directory.

This fixes the problem by keeping the old field value when the
ldap_email() or ldap_realname_from_username() functions return false.

Fixes #24720