Update Securimage library to custom 3.6.8 version
The 'mantis' branch of our Securimage fork [1] was updated to fix
PHP 8.1 compatibility issues. The initial patch was submitted as a PR to
the upstream repository [2], and a few additional fixes covering the
audio-related scripts were committed and submitted to the PR's author.

This commit updates the composer lock file, to pull these changes in.

Fixes #30794

[1]: https://github.com/mantisbt/securimage
[2]: https://github.com/dapphp/securimage/pull/112