Implementing the Status field of a category
The status field in the Category table is present for quite some time
but not being used.

Categories can be deactivated, meaning that they will no longer be
available when creating or updating an issue. Existing issues will keep
their current category even if it is deactivated.

Fixes #31017

This is the original contribution by @cas from the pull request
https://github.com/mantisbt/mantisbt/pull/1853, with all subsequent
corrective commits squashed, and commit message reformatted.

Signed-off-by: Damien Regad <dregad@mantisbt.org>