New bootstrap variable MANTIS_TESTSUITE_MAX_ISSUES
This replaces the FilterTest::ISSUES_TO_RETRIEVE constant, allowing a
developer to adjust the number of issues to retrieve when executing
FilterTest cases.

This avoids skipped tests when using a local database with more Issues
than the default of 50 (which works fine for TravisCI tests, that start
with an empty DB).

This is a follow-up fix for Issue #17121.