Define the Test Suites with a phpunit.xml file
Until today, the test suites were defined in "AllTests.php" files using
the PHPUnit\Framework\TestSuite class, which is meant for internal use.

The official, documented method to compose a test suite is to use XML
configuration.

This commit introduces the phpunit.xml, which replaces the AllTests.php
scripts, and updates documentation and TravisCI script.

Fixes #32816