Exception if relationship type id does not exist
IssueRelationshipAddCommand checks for invalid relationship name in the
payload, but lacks the same check when using a relationship id.

This lets the API create invalid relationships that can't be deleted and
cause errors when viewing the related Issues in the GUI.

Also changed the error code used for invalid rel type, to make it
consistent with invalid rel name.

Fixes #32835