Bad timestamp in VersionGetCommand::VersionToArray()
Missing ApiObjectFactory::$soap = false initialization causes the
version's date_order to be initialized as a XSD_DATETIME SoapVar instead
of an ISO-formatted date string as expected.

As long as the PHP SOAP Extension is available, this does not cause any
actual side effects, because
- the return value is not used when the Commands are used from within
  Mantis Core
- REST API initializes ApiObjectFactory::$soap = false so date_order is
  returned in the correct format.

If the SOAP extension is not available, Mantis triggers an internal
application error when this function is called.

Fixes #32909, #30415