Continue admin checks after failed email checks
Issue #32787 introduces a new check for email addresses that can be hard
and time-consuming to fix when failing.

This change allows to run the remaining tests even if the email check
failed.

Signed-off-by: Damien Regad <dregad@mantisbt.org>

This is a temporary workaround that should be reverted when implementing
the fix for Issue #33012. A @TODO comment was added in the code as a
reminder to do it when implementing a proper fix.

Initially submitted in PR https://github.com/mantisbt/mantisbt/pull/1928
(original commit message modified).