Made category default error message more generic
Originally, the message was only about deleting a category that is set
as default for moves. This makes the message more generic, so it can
also be used in other scenarios such as update.
Made category default error message more generic

Originally, the message was only about deleting a category that is set
as default for moves. This makes the message more generic, so it can
also be used in other scenarios such as update.

The ERROR_CATEGORY_CANNOT_DELETE_DEFAULT constant was renamed to
ERROR_CATEGORY_CANNOT_UPDATE_DEFAULT to reflect the new semantics.

Issue #31017