Move buttons to Edit User section footer
Until now, Reset Password, Delete User and Impersonate User buttons were
displayed between the Edit User and Add user to project sections.

This was dictated by legacy HTML limitation where it was not possible to
have multiple buttons in a single form with different submit actions.

With HTML5 we can leverage the formaction attribute, and move the
buttons into the Edit User section's footer for a cleaner UI.

Fixes #33842