Remove inclusion of core.php in timeline_inc.php
By definition an include file is part of a main script, which will
always include core.php, so it's not necessary to reference it again
in the include.

This actually caused an error on a user's system (PHP 8.2 on IIS). The
problem could not be reproduced, but an easy fix is to remove the
useless require_once.

Fixes #33906