Optimize create_cumulative_bydate() function
Until today, the function was executing simple SQL queries returning
atomic data, and PHP code had to loop over each row to compute the total
number of reported and resolved Issues.

This reports the effort to aggregate the data on the database server,
greatly simplifying the PHP code

It also fixes a bug where the last resolved issue was sometimes not
computed, resulting in an incorrect result set.

Fixes #34041,#34042