Escape label in print_link()
As suggested by @atrol, it makes sense to escape the string to output
at the lowest possible level, just before it is printed. This reduces
the risk of forgetting to do it in the parent function.

All usages of print_link() have been reviewed to ensure we do not have
double-escaping following this change.