Make "overdue" column sortable
Until now, sorting the Issues list by clicking on the overdue column
header only worked if the due_date column was also visible.

To fix the problem, we add due-date to the list of visible columns in
filter_get_visible_sort_properties_array() function if it is not already
part of the list.

Fixes #34460