Dropzone: adapt CSS to show errors
Until now it was difficult for users to understand that a file was
not successfully dropped, requiring use to hover the mouse over the
preview box, with no visible clue that something went wrong.

Errors occurring when adding a file are now clearly visible, with a
light red background on the preview box, and a bright red error icon
on with the error message displayed permanently next to it.

Includes minor design adjustments, e.g. truncating file name if longer
than preview box's maximum width

Finally, the CSS was refactored to remove unused rules and regroup
duplicate styles.

Fixes #34464