Remove code duplication in check_api.php
Functions check_print_test_row() and check_print_test_warn_row() were
almost identical, the only difference being in the logic to determine
if the check is a warning or failure.

A new optional parameter $p_warning was added to check_print_test_row()
and the logic adapted, allowing removal of duplicat code in
check_print_test_warn_row() which is reduce to a simple call to
check_print_test_row().

Fixes #34492