Move fastcgi_finish_request() to email shutdown function
This was originally added to html_end() to improve performance when
using php-fpm and sending mail synchronously [1], back when the function
was calling email_send_all(), i.e. before issue #17460 moved that to
email_shutdown_function().

This is now causing issues as the HTTP response code is not set when an
error occurs and FastCGI is used.

Fixes #34828, #34634

[1]: see commit cea405ccf228fd2c6ac694574a74e87396b14f1f