Use a stacked y axis for Issue Trend chart
The modified visualization introduced in commit  fa96f4399fc9c2f84bba815fcc230585087af525
was confusing, due to the `Still open` dataset being the delta of
`Reported` - `Resolved`.

We now display the areas under the `resolved` and `still open` datasets
as stacked data. Since the sum of the 2 is equal to the `reported` set,
its line overlaps and we avoid confusing display.

Fixes #34847