Ensure user in manage_config_cookie exists
If a non-existing user id is specified in the cookie, user_get_row()
returns false, resulting in the subsequent user_get_name_from_row() call
triggering a PHP error as the function expects an array but receives a
bool.

Adding a check for the user id's existence after reading the cookie and
defaulting to ALL_USERS if not found fixes the problem.

Fixes #35064