Allow delaying of warning messages
Warnings triggered by MantisBT core before the page header has been
printed cause the layout to be broken if warnings are configured to be
displayed inline.

Introduce a new error_delay_reporting() function, which allows to
selectively enable the delayed printing of warnings (by defining the
DELAY_INLINE_ERROR_REPORTING constant), and use it in event_hook() and
in the error_handler() for the E_USER_DEPRECATED error type.

Adapt the error handler to delay inline warnings when the constant is
defined.

Fixes #35207