Fix Bug Report Page redirect behavior
If an early error occurs when calling bug_report_page.php with
ALL_PROJECTS as the current project, a redirection occurs but since
handled errors have been logged print_header_redirect() displays a
"Proceed" message so the inline errors can be shown, but does not
interrupt the script, causing the page layout to be printed twice and
cascading errors to occur.

Testing print_header_redirect()'s return value and aborting the script
if false fixes the problem.

Issue ##35208