Restore included pages functionality
- Restored working of top_include_page and bottom_include_page options.
- Restored generation of EVENT_LAYOUT_PAGE_HEADER event.
- Added dynamic footer resizing.

The top_include_page file is included within the navbar element, which
is fixed at the top of the page; the bottom_include_page file is
included within an existing footer or as a footer itself (on login
pages), which is also fixed at the bottom of the page. Because of this,
the include template must contain, for example, a “div” element with a
solid background.

Fixes #35402, Fixes #22250, Fixes #22098.