Prevent internal application error in url_get()
When allow_url_fopen = OFF, curl extension is not installed, url_get()
falls back to a shell_exec() call to run curl, but if shell_exec()
function is disabled in php.ini, then an internal application error is
thrown.

This checks that shell_exec() function exists before calling it, and
returns null if not.

Fixes #35540