Fix scroll offset on bugnote links
When clicking on a direct link to a bugnote, e.g. from "View Issue
Details" page or from a notification email, MantisBT did not account for
the Navbar's height, causing the upper part of the note to be hidden
behind it.

Adding scroll-margin-top CSS attribute allows proper positioning.

Fixes #21675

Signed-off-by: Damien Regad <dregad@mantisbt.org>

Changes to original contribution from [PR 2137][1]
- Rewritten commit message, added bugtracker reference
- Coding guidelines

[1]: https://github.com/mantisbt/mantisbt/pull/2137