Fix additional scroll offset issues
In most use cases, we want to jump to a section (i.e. a div with the
widget-box class); many of these already have an id, so we just need to
extend the scroll offset rule by adding the `div.widget-box` selector
to fix those.

Additionally, this removes several unnecessary <a> tags that are
duplicating existing IDs on <div> with widget-box class.

- View page:
  - Monitors section (#monitors)
  - Notes/Activities section (#bugnotes, #attachments)
  - Time tracking section (#bugnotestats)
  - History section (#history)
- Manage Tags page, Create Tag section (#tagcreate)
- Configuration Report page, Database Configuration section
- Language tests page, Plugins checks (add `.test-langs tr` selector)

Fixes #21675