Allow deleting a Global config option
config_delete() prevents deleting configs defined in $g_global_settings.
This does not make sense: if they are not allowed to be present in the
database, then it should be OK to remove them.

This situation can happen, when a config that was not initially defined
as global, is later added to the $g_global_settings.

Fixes #36164