Prevent HTML output in Dropzone when errors occur
A somewhat hackish solution, leveraging the existing behavior in the
error handler, driven by the DISABLE_INLINE_ERROR_REPORTING constant.

When bug_report.php or bugnote_add.php are called via AJAX (as is the
case when Dropzone handles file uploads), the constant is set with a new
value of `text`. This is picked up by error_handler(), which sends the
error message in the response body as plain text.

Note: AJAX detection relies on presence of non-standard X-Requested-With
header with value 'XMLHttpRequest' (which is set by Dropzone [1]).

Fixes #36303

[1]: https://github.com/mantisbt/mantisbt/blob/38a569855832c9d5cc63ee8bdf5a96c9b0c60e84/js/dropzone-5.5.0.js#L2446